Rename `embed` to `ndlaembed`
This comes with a small gotcha that every `ndlaembed` now needs a
closing tag.

Html only supports self-closing tags for void elements (which `embed`
was by luck), but now that we want a custom tag we need to include a
closing tag manually.

This looks kind of weird, but it works so lets roll with it.